Moved foreach_handler to a dedicated dir
