+++ checkov/terraform/graph_builder/foreach/handler.py
