+++ checkov/terraform/graph_builder/local_graph.py
@@ -20,7 +20,7 @@ from checkov.common.util.type_forcers import force_int
 from checkov.terraform.modules.module_objects import TFModule
 from checkov.terraform.checks.utils.dependency_path_handler import unify_dependency_path
 from checkov.terraform.context_parsers.registry import parser_registry
-import checkov.terraform.graph_builder.foreach_handler as foreach_module
+import checkov.terraform.graph_builder.foreach.handler as foreach_module
 from checkov.terraform.graph_builder.graph_components.block_types import BlockType
 from checkov.terraform.graph_builder.graph_components.blocks import TerraformBlock
 from checkov.terraform.graph_builder.graph_components.generic_resource_encryption import ENCRYPTION_BY_RESOURCE_TYPE
